refactor(frontend): move Nav component to elements folder
