+++ packages/frontend/src/components/containers/index.ts
@@ -1,10 +1,9 @@
 import RegisterForm from './RegisterForm';
 import LoginForm from './LoginForm';
 import Products from './Products';
-import Nav from './Nav';
 import PayPal from './PayPal';
 import CartContent from './CartContent';
 import UserForms from './UserForms';
 import UserProducts from './UserProducts';
 
-export { RegisterForm, LoginForm, Products, Nav, PayPal, CartContent, UserForms, UserProducts };
+export { RegisterForm, LoginForm, Products, PayPal, CartContent, UserForms, UserProducts };
