+++ packages/frontend/src/components/elements/index.ts
@@ -1,0 +1,3 @@
+import Nav from './Nav';
+
+export { Nav };
